refactor(system): 优化系统模块代码结构和错误处理
- 修改了学生、教师、权限、角色和用户相关服务的错误处理逻辑
- 使用枚举类统一错误码和错误信息
- 优化了数据库表结构，将 int 类型改为 tinyint(1)以节省空间
- 修复了一些潜在的 bug 和性能问题

+++ init/mysql/system.sql
@@ -11,7 +11,7 @@ create table sys_permission
     sort            int      default 0                 null,
     create_time     datetime default CURRENT_TIMESTAMP null,
     update_time     datetime default CURRENT_TIMESTAMP null on update CURRENT_TIMESTAMP,
-    is_deleted      int      default 0                 null
+    is_deleted      tinyint(1)      default 0          null
 );
 
 -- ===============================================================
@@ -24,11 +24,11 @@ create table sys_role
     role_name   varchar(30)                        not null,
     role_key    varchar(100)                       not null,
     sort        int      default 0                 null,
-    status      int      default 0                 null,
+    status      tinyint(1)      default 0          null,
     description varchar(500)                       null,
     create_time datetime default CURRENT_TIMESTAMP null,
     update_time datetime default CURRENT_TIMESTAMP null on update CURRENT_TIMESTAMP,
-    is_deleted  int      default 0                 null
+    is_deleted  tinyint(1)      default 0          null
 );
 
 -- ===============================================================
@@ -43,13 +43,13 @@ create table sys_user
     nick_name       varchar(255)                       null,
     email           varchar(255)                       null,
     mobile          varchar(255)                       null,
-    gender          int                                null,
+    gender          tinyint(1)                         null,
     avatar          varchar(255)                       null,
-    status          int      default 0                 null,
+    status          tinyint(1)      default 0          null,
     last_login_time datetime                           null,
     create_time     datetime default CURRENT_TIMESTAMP null,
     update_time     datetime default CURRENT_TIMESTAMP null on update CURRENT_TIMESTAMP,
-    is_deleted      int      default 0                 null
+    is_deleted      tinyint(1)      default 0          null
 );
 
 -- ===============================================================
